construct M-SEARCH packet from template string in PROGMEM to save RAM

+++ UPnP/Discover.h
@@ -26,7 +26,7 @@ public:
 	// for every response received within the timeout, the callback is invoked
 	// if the callback returns false, discovery is terminated
 	// returns false if discovery failed; not receiving any response is NOT considered a failure
-	bool discover(DiscoverCallback callback, const char *st, int mx = 4, unsigned long timeoutMillis = 5000);
+	bool discover(DiscoverCallback callback, const char *st, uint8_t mx = 4, unsigned long timeoutMillis = 5000);
 };
 
 }
